Integrated Java Program
Updated java program to be able to connect to the gradebook database and either view a list of students and their grades in a specific function or a list of sections of which a student is enrolled in, along with their grade. Queries will possibly need an update to work properly.